Project/requirements.txt
flask
flask-cors
PyJWT
bcrypt
pymysql
cryptography
